Title: 최솟값 3개, Time: 180ms, Memory: 1MB, Status: Passed - CodeTree

+++ 240625/최솟값 3개/top-3-smallest-number.cpp
@@ -27,7 +27,7 @@ int main() {
             c = pq.top();
             pq.pop();
 
-            cout << a * b * c << endl;
+            cout << (long long)a * b * c << endl;
 
             pq.push(a);
             pq.push(b);
